pyproject.toml
[tool.poetry]
name = "django_starterkit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^2.0.13"
psycopg2-binary = "^2.7.7"
djangorestframework = "^3.10.2"
django-filter = "^2.2"
celery = "^4.3.0"
redis = "^3.3.6"
django-extensions = "^2.2"
drf-yasg = "^1.17"

[tool.poetry.dev-dependencies]
pylint = "^2.3"
coverage = "^4.5.1"
pylint-celery = "^0.3"
pylint-django = "^2.0.11"
bandit = "^1.6.0"
radon = "^3.0.3"
Faker = "^2.0.1"
pytest = "^5.1"
pytest-django = "^3.5"
pytest-mock = "^1.10"
model_mommy = "^1.6"
freezegun = "^0.3.12"
pytest-cov = "^2.7"
dj-database-url = "^0.5.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
